feat(openapi): add fields for number of threads for Dask workers (#484)
Following the integration of Dask into REANA and discussions with
several REANA users, we identified the need to make the number of
threads configurable.

This commit introduces an additional field to allow users to specify
the number of threads. The field is optional, similar to other Dask
parameters, except for the image field.

The OpenAPI specification for the `api/info` command has been updated
to include the default and maximum number of threads option for Dask
workers.

Closes reanahub/reana#874